+++ src/duckdb/duckdb_server.cpp
@@ -1166,6 +1166,7 @@ class DuckDBFlightSqlServer::Impl {
     max_sessions_ = max_sessions;
     session_idle_timeout_seconds_ = session_idle_timeout_seconds;
     StartIdleSessionSweeper();
+    ::gizmosql::metrics::RegisterAdmissionQueueGauges(admission_controller_);
   }
 
   std::shared_ptr<InstrumentationManager> GetInstrumentationManager() const {
@@ -1210,6 +1211,7 @@ class DuckDBFlightSqlServer::Impl {
     max_sessions_ = max_sessions;
     session_idle_timeout_seconds_ = session_idle_timeout_seconds;
     StartIdleSessionSweeper();
+    ::gizmosql::metrics::RegisterAdmissionQueueGauges(admission_controller_);
   }
 #endif